--- a/pkg/ccl/changefeedccl/kvevent/BUILD.bazel
+++ b/pkg/ccl/changefeedccl/kvevent/BUILD.bazel
@@ -36,15 +36,18 @@ go_test(
     name = "kvevent_test",
     srcs = [
         "alloc_test.go",
+        "bench_test.go",
         "blocking_buffer_test.go",
     ],
     embed = [":kvevent"],
     deps = [
+        "//pkg/jobs/jobspb",
         "//pkg/keys",
         "//pkg/roachpb",
         "//pkg/settings/cluster",
         "//pkg/sql/randgen",
         "//pkg/sql/rowenc/keyside",
+        "//pkg/sql/sem/tree",
         "//pkg/sql/types",
         "//pkg/util/ctxgroup",
         "//pkg/util/encoding",