This change removes a pointer from the kvevent.Event struct, reducing overall allocations. The hope is that this reduces the amount of work Go gc has to do, which will reduce SQL latency at the end of the day. When doing backfills, the allocations in kv events add up pretty fast, so reducing even one pointer is significant.

See https://github.com/cockroachdb/cockroach/issues/84709 for more info. I'm not closing the issue with this PR since we may decide to reduce more pointers in future PRs using some of the ideas in the issue comments.

Here are the benchmark results 
```
name          old time/op    new time/op    delta
MemBuffer-10    98.1µs ± 0%    95.8µs ± 1%   -2.35%  (p=0.008 n=5+5)

name          old alloc/op   new alloc/op   delta
MemBuffer-10    76.9kB ± 0%    64.4kB ± 0%  -16.17%  (p=0.008 n=5+5)

name          old allocs/op  new allocs/op  delta
MemBuffer-10       859 ± 0%       675 ± 0%  -21.42%  (p=0.008 n=5+5)

```