@miretskiy 2022-07-28
Is the delta not shown in your PR description because you haven't run this enough times?
Perhaps run the benchmark at least 10 times to see delta %-ges?