@jayshrivastava 2022-11-10
@shermanCRL I don't think we should do it. The benefit is very small. In https://github.com/cockroachdb/cockroach/pull/86421, Nathan mentioned this may have no benefit in some scenarios.